Fixed up the file structure. Removed THREAD_SAFE define... user should provide own thread-safety mechanisms.
