--- Makefile
@@ -1,235 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 2.8
-
-# Default target executed when no arguments are given to make.
-default_target: all
-.PHONY : default_target
-
-#=============================================================================
-# Special targets provided by cmake.
-
-# Disable implicit rules so canonical targets will work.
-.SUFFIXES:
-
-# Remove some rules from gmake that .SUFFIXES does not remove.
-SUFFIXES =
-
-.SUFFIXES: .hpux_make_needs_suffix_list
-
-# Suppress display of executed commands.
-$(VERBOSE).SILENT:
-
-# A target that is always out of date.
-cmake_force:
-.PHONY : cmake_force
-
-#=============================================================================
-# Set environment variables for the build.
-
-# The shell in which to execute make rules.
-SHELL = /bin/sh
-
-# The CMake executable.
-CMAKE_COMMAND = /usr/bin/cmake
-
-# The command to remove a file.
-RM = /usr/bin/cmake -E remove -f
-
-# Escaping for special characters.
-EQUALS = =
-
-# The top-level source directory on which CMake was run.
-CMAKE_SOURCE_DIR = /home/brian/Repo/xml
-
-# The top-level build directory on which CMake was run.
-CMAKE_BINARY_DIR = /home/brian/Repo/xml
-
-#=============================================================================
-# Targets provided globally by CMake.
-
-# Special rule for the target edit_cache
-edit_cache:
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --cyan "Running interactive CMake command-line interface..."
-	/usr/bin/cmake -i .
-.PHONY : edit_cache
-
-# Special rule for the target edit_cache
-edit_cache/fast: edit_cache
-.PHONY : edit_cache/fast
-
-# Special rule for the target install
-install: preinstall
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --cyan "Install the project..."
-	/usr/bin/cmake -P cmake_install.cmake
-.PHONY : install
-
-# Special rule for the target install
-install/fast: preinstall/fast
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --cyan "Install the project..."
-	/usr/bin/cmake -P cmake_install.cmake
-.PHONY : install/fast
-
-# Special rule for the target install/local
-install/local: preinstall
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --cyan "Installing only the local directory..."
-	/usr/bin/cmake -DCMAKE_INSTALL_LOCAL_ONLY=1 -P cmake_install.cmake
-.PHONY : install/local
-
-# Special rule for the target install/local
-install/local/fast: install/local
-.PHONY : install/local/fast
-
-# Special rule for the target install/strip
-install/strip: preinstall
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --cyan "Installing the project stripped..."
-	/usr/bin/cmake -DCMAKE_INSTALL_DO_STRIP=1 -P cmake_install.cmake
-.PHONY : install/strip
-
-# Special rule for the target install/strip
-install/strip/fast: install/strip
-.PHONY : install/strip/fast
-
-# Special rule for the target list_install_components
-list_install_components:
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --cyan "Available install components are: \"Unspecified\""
-.PHONY : list_install_components
-
-# Special rule for the target list_install_components
-list_install_components/fast: list_install_components
-.PHONY : list_install_components/fast
-
-# Special rule for the target rebuild_cache
-rebuild_cache:
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --cyan "Running CMake to regenerate build system..."
-	/usr/bin/cmake -H$(CMAKE_SOURCE_DIR) -B$(CMAKE_BINARY_DIR)
-.PHONY : rebuild_cache
-
-# Special rule for the target rebuild_cache
-rebuild_cache/fast: rebuild_cache
-.PHONY : rebuild_cache/fast
-
-# The main all target
-all: cmake_check_build_system
-	$(CMAKE_COMMAND) -E cmake_progress_start /home/brian/Repo/xml/CMakeFiles /home/brian/Repo/xml/CMakeFiles/progress.marks
-	$(MAKE) -f CMakeFiles/Makefile2 all
-	$(CMAKE_COMMAND) -E cmake_progress_start /home/brian/Repo/xml/CMakeFiles 0
-.PHONY : all
-
-# The main clean target
-clean:
-	$(MAKE) -f CMakeFiles/Makefile2 clean
-.PHONY : clean
-
-# The main clean target
-clean/fast: clean
-.PHONY : clean/fast
-
-# Prepare targets for installation.
-preinstall: all
-	$(MAKE) -f CMakeFiles/Makefile2 preinstall
-.PHONY : preinstall
-
-# Prepare targets for installation.
-preinstall/fast:
-	$(MAKE) -f CMakeFiles/Makefile2 preinstall
-.PHONY : preinstall/fast
-
-# clear depends
-depend:
-	$(CMAKE_COMMAND) -H$(CMAKE_SOURCE_DIR) -B$(CMAKE_BINARY_DIR) --check-build-system CMakeFiles/Makefile.cmake 1
-.PHONY : depend
-
-#=============================================================================
-# Target rules for targets named XMLTest
-
-# Build rule for target.
-XMLTest: cmake_check_build_system
-	$(MAKE) -f CMakeFiles/Makefile2 XMLTest
-.PHONY : XMLTest
-
-# fast build rule for target.
-XMLTest/fast:
-	$(MAKE) -f CMakeFiles/XMLTest.dir/build.make CMakeFiles/XMLTest.dir/build
-.PHONY : XMLTest/fast
-
-source/xml.o: source/xml.cpp.o
-.PHONY : source/xml.o
-
-# target to build an object file
-source/xml.cpp.o:
-	$(MAKE) -f CMakeFiles/XMLTest.dir/build.make CMakeFiles/XMLTest.dir/source/xml.cpp.o
-.PHONY : source/xml.cpp.o
-
-source/xml.i: source/xml.cpp.i
-.PHONY : source/xml.i
-
-# target to preprocess a source file
-source/xml.cpp.i:
-	$(MAKE) -f CMakeFiles/XMLTest.dir/build.make CMakeFiles/XMLTest.dir/source/xml.cpp.i
-.PHONY : source/xml.cpp.i
-
-source/xml.s: source/xml.cpp.s
-.PHONY : source/xml.s
-
-# target to generate assembly for a file
-source/xml.cpp.s:
-	$(MAKE) -f CMakeFiles/XMLTest.dir/build.make CMakeFiles/XMLTest.dir/source/xml.cpp.s
-.PHONY : source/xml.cpp.s
-
-xml_test.o: xml_test.cpp.o
-.PHONY : xml_test.o
-
-# target to build an object file
-xml_test.cpp.o:
-	$(MAKE) -f CMakeFiles/XMLTest.dir/build.make CMakeFiles/XMLTest.dir/xml_test.cpp.o
-.PHONY : xml_test.cpp.o
-
-xml_test.i: xml_test.cpp.i
-.PHONY : xml_test.i
-
-# target to preprocess a source file
-xml_test.cpp.i:
-	$(MAKE) -f CMakeFiles/XMLTest.dir/build.make CMakeFiles/XMLTest.dir/xml_test.cpp.i
-.PHONY : xml_test.cpp.i
-
-xml_test.s: xml_test.cpp.s
-.PHONY : xml_test.s
-
-# target to generate assembly for a file
-xml_test.cpp.s:
-	$(MAKE) -f CMakeFiles/XMLTest.dir/build.make CMakeFiles/XMLTest.dir/xml_test.cpp.s
-.PHONY : xml_test.cpp.s
-
-# Help Target
-help:
-	@echo "The following are some of the valid targets for this Makefile:"
-	@echo "... all (the default if no target is provided)"
-	@echo "... clean"
-	@echo "... depend"
-	@echo "... XMLTest"
-	@echo "... edit_cache"
-	@echo "... install"
-	@echo "... install/local"
-	@echo "... install/strip"
-	@echo "... list_install_components"
-	@echo "... rebuild_cache"
-	@echo "... source/xml.o"
-	@echo "... source/xml.i"
-	@echo "... source/xml.s"
-	@echo "... xml_test.o"
-	@echo "... xml_test.i"
-	@echo "... xml_test.s"
-.PHONY : help
-
-
-
-#=============================================================================
-# Special targets to cleanup operation of make.
-
-# Special rule to run CMake to check the build system integrity.
-# No rule that depends on this can have commands that come from listfiles
-# because they might be regenerated.
-cmake_check_build_system:
-	$(CMAKE_COMMAND) -H$(CMAKE_SOURCE_DIR) -B$(CMAKE_BINARY_DIR) --check-build-system CMakeFiles/Makefile.cmake 0
-.PHONY : cmake_check_build_system
-
