--- source/CMakeFiles/progress.marks
@@ -1,1 +1,0 @@
-0
