+++ test/xml_test.cpp
@@ -1,18 +1,22 @@
 /// XML Test
-
-#include "xml.hpp"
+#include <iostream>
 #include <cmath>
+#include <string>
 
+#include <xml/xml.h>
 
 int main(int argc, char** argv )
 {
+	std::string fname = "xml_output_test.xml";
+
 	/// Writer Test
-	std::ofstream test_fos("xml_output_test.xml");
+	std::cout << "Creating an XML document" << std::endl;
+	std::ofstream test_fos(fname.c_str());
 
 	XML::writer	wr(test_fos);
 
-	wr.beg("doc",1)
-		.set_param<const char*>("name","test",XML::writer::INLINE);
+	wr.beg("doc",1UL)
+		.set_param<std::string>("name", "test", XML::writer::INLINE);
 		wr.beg("d",3UL)
 			.set_param<float>("norm",	1.0f/sqrtf(2.0f))
 			.set_param<float>("X",		1.0f)
@@ -23,12 +27,12 @@ int main(int argc, char** argv )
 		wr.end("d");
 
 	test_fos.close();
-
+	std::cout << "Done creating an XML document" << std::endl;
 
 
 	/// Reader Test
-	std::ifstream 	test_fis("xml_output_test.xml");
-	std::string 	sval;
+	std::ifstream 	test_fis(fname.c_str());
+	std::string sval;
 	float 			norm;
 	float 			X;
 	float 			Y;
@@ -36,7 +40,8 @@ int main(int argc, char** argv )
 	XML::reader	rd(test_fis);
 
 	//<doc ... />
-	rd.next("doc").get_param<std::string>("name",sval);
+	rd.next("doc")
+		.get_param<std::string>("name",sval);
 
 	//<d>
 	rd.next("d")
@@ -47,6 +52,7 @@ int main(int argc, char** argv )
 		//<p> ... </p>
 		rd.into("p").get_content(std::cout);
 	//</d>
+
 
 
 	std::cout 	<< sval
